+++ lua/obsidian/config/default.lua
@@ -1,22 +1,53 @@
 ---@type obsidian.config.Internal
 return {
-  legacy_commands = true, -- TODO: remove in 4.0.0
-  workspaces = {},
-  log_level = vim.log.levels.INFO,
+  -- TODO: remove these in 4.0.0
+  legacy_commands = true,
+  note_frontmatter_func = require("obsidian.builtin").frontmatter,
+  disable_frontmatter = false,
+  ---@class obsidian.config.StatuslineOpts
+  ---
+  ---@field format? string
+  ---@field enabled? boolean
+  statusline = {
+    format = "{{backlinks}} backlinks  {{properties}} properties  {{words}} words  {{chars}} chars",
+    enabled = true,
+  },
+
+  -- TODO:: replace with more general options before 4.0.0
+  follow_url_func = vim.ui.open,
+  follow_img_func = vim.ui.open,
   notes_subdir = nil,
   new_notes_location = "current_dir",
+
+  -- TODO: group into a search module
+  sort_by = "modified",
+  sort_reversed = true,
+  search_max_lines = 1000,
+
+  workspaces = {},
+  log_level = vim.log.levels.INFO,
   note_id_func = require("obsidian.builtin").zettel_id,
   wiki_link_func = require("obsidian.builtin").wiki_link_id_prefix,
   markdown_link_func = require("obsidian.builtin").markdown_link,
   preferred_link_style = "wiki",
-  follow_url_func = vim.ui.open,
-  follow_img_func = vim.ui.open,
-  note_frontmatter_func = require("obsidian.builtin").frontmatter,
-  disable_frontmatter = false,
-  sort_by = "modified",
-  sort_reversed = true,
-  search_max_lines = 1000,
   open_notes_in = "current",
+
+  ---@class obsidian.config.FrontmatterOpts
+  ---
+  --- Whether to enable frontmatter, boolean for global on/off, or a function that takes filename and returns boolean.
+  ---@field enabled? (fun(fname: string?): boolean)|boolean
+  ---
+  --- Function to turn Note attributes into frontmatter.
+  ---@field func? fun(note: obsidian.Note): table<string, any>
+  --- Function that is passed to table.sort to sort the properties, or a fixed order of properties.
+  ---
+  --- List of string that sorts frontmatter properties, or a function that compares two values, set to vim.NIL/false to do no sorting
+  ---@field sort? string[] | (fun(a: any, b: any): boolean) | vim.NIL | boolean
+  frontmatter = {
+    enabled = true,
+    func = require("obsidian.builtin").frontmatter,
+    sort = { "id", "aliases", "tags" },
+  },
 
   ---@class obsidian.config.TemplateOpts
   ---
@@ -207,15 +238,6 @@ return {
   ---Runs anytime the workspace is set/changed.
   ---@field post_set_workspace? fun(workspace: obsidian.Workspace)
   callbacks = {},
-
-  ---@class obsidian.config.StatuslineOpts
-  ---
-  ---@field format? string
-  ---@field enabled? boolean
-  statusline = {
-    format = "{{backlinks}} backlinks  {{properties}} properties  {{words}} words  {{chars}} chars",
-    enabled = true,
-  },
 
   ---@class obsidian.config.FooterOpts
   ---