+++ lua/obsidian/types.lua
@@ -47,6 +47,7 @@
 ---@field follow_img_func? fun(img: string)
 ---@field note_frontmatter_func? (fun(note: obsidian.Note): table)
 ---@field disable_frontmatter? (fun(fname: string?): boolean)|boolean
+---@field frontmatter? obsidian.config.FrontmatterOpts
 ---@field backlinks? obsidian.config.BacklinkOpts
 ---@field completion? obsidian.config.CompletionOpts
 ---@field picker? obsidian.config.PickerOpts
@@ -79,8 +80,7 @@
 ---@field preferred_link_style obsidian.config.LinkStyle
 ---@field follow_url_func fun(url: string)|?
 ---@field follow_img_func fun(img: string)|?
----@field note_frontmatter_func (fun(note: obsidian.Note): table)|?
----@field disable_frontmatter (fun(fname: string?): boolean)|boolean|?
+---@field frontmatter obsidian.config.FrontmatterOpts
 ---@field backlinks obsidian.config.BacklinkOpts
 ---@field completion obsidian.config.CompletionOpts
 ---@field picker obsidian.config.PickerOpts