+++ lua/obsidian/yaml/init.lua
@@ -6,6 +6,7 @@ local yaml = {}
 ---Deserialize a YAML string.
 ---@param str string
 ---@return any
+---@return string[]
 yaml.loads = function(str)
   return parser.loads(str)
 end
@@ -95,7 +96,7 @@ end
 
 ---Dump an object to YAML lines.
 ---@param x any
----@param order function
+---@param order function?
 ---@return string[]
 yaml.dumps_lines = function(x, order)
   return dumps(x, 0, order)