+++ tests/lsp/test_rename.lua
@@ -1,25 +1,21 @@
-local new_set, eq = MiniTest.new_set, MiniTest.expect.equality
-local child = MiniTest.new_child_neovim()
+local eq = MiniTest.expect.equality
 local Path = require "obsidian.path"
+local h = dofile "tests/helpers.lua"
 
-local T = new_set {
-  hooks = {
-    pre_case = function()
-      child.restart { "-u", "scripts/minimal_init_with_setup.lua" }
-      child.lua [[
-Note = require"obsidian.note"
-Obsidian.opts.disable_frontmatter = true
-      ]]
-    end,
-    post_once = function()
-      child.lua [[vim.fn.delete(tostring(Obsidian.dir), "rf")]]
-      child.stop()
-    end,
-  },
-}
+local T, child = h.child_vault()
 
 local target = [[---
 id: target
+aliases: []
+tags: []
+---
+hello
+world]]
+
+local target_expected = [[---
+id: new_target
+aliases: []
+tags: []
 ---
 hello
 world]]
@@ -36,18 +32,14 @@ end
 T["rename current note"] = function()
   local root = Path.new(child.lua_get [[tostring(Obsidian.dir)]])
   local target_path = root / "target.md"
-  local referencer_path = root / "referencer.md"
   write(target, target_path)
-  write(referencer, referencer_path)
 
   child.lua(string.format([[vim.cmd("edit %s")]], target_path))
   child.lua [[vim.lsp.buf.rename("new_target", {})]]
   local new_target_path = root / "new_target.md"
   eq(true, new_target_path:exists())
-  local bufs = child.api.nvim_list_bufs()
-  eq(3, #bufs)
   local lines = child.api.nvim_buf_get_lines(1, 0, -1, false) -- new_target
-  eq("id: new_target", lines[2])
+  eq(target_expected, table.concat(lines, "\n"))
 end
 
 T["rename note under cursor"] = function()
@@ -66,12 +58,22 @@ T["rename note under cursor"] = function()
   local new_target_path = root / "new_target.md"
   local lines = vim.fn.readfile(tostring(new_target_path))
 
-  eq("id: new_target", lines[2])
+  eq(target_expected, table.concat(lines, "\n"))
 end
 
 local referencer2 = [==[
 
 [[target#^block]]
+]==]
+
+local referencer2_expected = [==[
+---
+id: referencer
+aliases: []
+tags: []
+---
+
+[[new_target#^block]]
 ]==]
 
 T["rename note without changing blocks and headers"] = function()
@@ -91,9 +93,9 @@ T["rename note without changing blocks and headers"] = function()
   local lines = vim.fn.readfile(tostring(new_target_path))
 
   local ref_lines = vim.fn.readfile(tostring(referencer_path))
-  eq([==[[[new_target#^block]]]==], ref_lines[2])
+  eq(referencer2_expected, table.concat(ref_lines, "\n"))
 
-  eq("id: new_target", lines[2])
+  eq(target_expected, table.concat(lines, "\n"))
 end
 
 return T